Replace dialog preferences altogether with simpler androidx edittexts
The dialog/preference paradigm was just a bit complex, with a lot of
supporting code. Replacing dialogs with simple preference edittexts and
groups is a more consistent and easy-to-use way of achieving the same
thing.

Also, some preferences are set to be their default on startup now if
they're a blank or empty value. Port, clientId, deviceId etc.